Fix problems with imports
* Fix them all
* Move `ScrollableFrame` to `widgets.py`
* Delete `scrollable_frame.py`
